Refactor DataTypes api
- Plugin now gets a pointer to its self, via mqplugin::ThisPlugin, and a
  pointer to the new MainInterface from MQ2Main
- Add implementations to pluginapi which will forward requests to MQ2Main
  over a new "Main" interface. This will include the plugin that made
  the request.
- Encapsulate the "DataAPI" into a class.
- Removed PlayerClient parameter from ParseMacroParameter (deprecated)
- Move more code out of MQ2Main.h and MQ2Internal.h etc into api headers.
- Partially rename some APIs (MQDataItem/MQ2Data -> TopLevelObject etc)
