+++ include/mq/plugin/pluginapi.h
@@ -16,6 +16,7 @@
 // not reside within MQ2Main.
 
 #include "../../src/main/MQ2Main.h"
+#include "mq/api/Main.h"
 
 #define PLUGIN_API extern "C" __declspec(dllexport)
 
@@ -26,6 +27,9 @@ namespace mqplugin {
 
 extern const char* PluginName;
 extern HINSTANCE ghPluginModule;
+extern mq::MainInterface* MainInterface;
+PLUGIN_API mq::MQPlugin* ThisPlugin;
+
 
 bool PluginMain(HINSTANCE hModule, DWORD dwReason, void* lpReserved);
 
@@ -43,4 +47,3 @@ bool PluginMain(HINSTANCE hModule, DWORD dwReason, void* lpReserved);
 #if __has_include("../../../src/private/pluginapi-private.h")
 #include "../../../src/private/pluginapi-private.h"
 #endif
-
